undo: Move command checkpoint handling under data.undo
Undo references, file capture, restoration, snapshot storage, and conflict
checks now live under data.undo. The context used around
repository-changing commands still shares data.undo_checkpoints with undo
and redo execution.

Commands open that context before a change so it can record the before
state and finalize the checkpoint after success. Those commands do not need
direct access to operations that move undo or redo stack pointers.

This commit moves command checkpoint creation and finalization into
data.undo.checkpoints and updates repository-changing commands to import it
there.

The next commit will update tests for the packaged command context before
redo and undo execution follow it.
