+++ src/git_stage_batch/commands/unblock_file.py
@@ -6,7 +6,7 @@ import sys
 from contextlib import nullcontext
 
 from ..data.session_marker import session_is_active
-from ..data.undo_checkpoints import undo_checkpoint
+from ..data.undo.checkpoints import undo_checkpoint
 from ..data.ignore_files import (
     add_pattern_to_gitignore,
     add_pattern_to_local_exclude,
